pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "commonbase"
description = "Commonbase Python SDK"
readme = "README.md"
license.file = "LICENSE"
requires-python = ">=3.8"
version = "0.1.2"
dependencies = ["requests>=2.31.0", "sseclient-py>=1.7.2"]
urls.Homepage = "https://github.com/commonbaseapp/commonbase-python"
urls.Source = "https://github.com/commonbaseapp/commonbase-python"
urls.Tracker = "https://github.com/commonbaseapp/commonbase-python/issues"
